requirement.txt
asgiref==3.11.0
bcrypt==5.0.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
PyJWT==2.10.1
pymongo==4.15.5
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
